✨ feature(logging): Multiple fixes after PR feedback
* Move structlog-specific initiation to `troncos/frameworks/` to decouple from non-structlog execution paths
* Put structlog dependency into an `extras` group. Installing with structlog support now requires the user to run `poetry install --extras "structlog"` or adding `troncos[structlog]` to their requirements
* Add `init_logging_structlog` as an entrypoint to handle logging initiation with structlog support
* Add `trace_injection_processor` to shared processor chain
* Add `merge_contextvars` to shared processor chain
* Add support for configuring structlog with extra processors
